+++ src/testmod/kotlin/me/fzzyhmstrs/fzzy_config_test/test/TestConfigImpl4.kt
@@ -12,8 +12,14 @@ package me.fzzyhmstrs.fzzy_config_test.test
 
 import me.fzzyhmstrs.fzzy_config.annotations.RequiresRestart
 import me.fzzyhmstrs.fzzy_config.config.Config
+import me.fzzyhmstrs.fzzy_config.validation.collection.ValidatedList
+import me.fzzyhmstrs.fzzy_config.validation.collection.ValidatedMap
+import me.fzzyhmstrs.fzzy_config.validation.collection.ValidatedStringMap
 import me.fzzyhmstrs.fzzy_config.validation.misc.ValidatedBoolean
+import me.fzzyhmstrs.fzzy_config.validation.misc.ValidatedString
+import me.fzzyhmstrs.fzzy_config.validation.number.ValidatedDouble
 import me.fzzyhmstrs.fzzy_config.validation.number.ValidatedInt
+import net.fabricmc.loader.api.FabricLoader
 import net.minecraft.util.Identifier
 
 @RequiresRestart
@@ -26,6 +32,12 @@ class TestConfigImpl4: Config(Identifier("fzzy_config_test","test_config4")) {
     @ValidatedInt.Restrict(0,20)
     var int1 = 6
     var int2 = ValidatedInt(6,10,1)
+
+    var mapDouble = ValidatedStringMap(mapOf("a" to 1.0),ValidatedString(), ValidatedDouble(1.0,1.0,0.0))
+
+    var namespaceBlackList: ValidatedList<String> = ValidatedString.fromList(FabricLoader.getInstance().allMods.map{ it.metadata.id }).toList()
+
+    var testString = ValidatedString.fromList(FabricLoader.getInstance().allMods.map{ it.metadata.id })
 
     /*
     {