+++ changelog.md
@@ -5,4 +5,6 @@
 * None.
 
 ### Fixes
-* Fixed new `ValidatedString` methods not returning a ValidatedString as intended.
+* Fixed blank line at top of description tooltips.
+* Fixed `MapListWidget` not properly firing `mouseReleased` on its children when the mouse is released out of that child's bounds.
+* Fixed suggestion lists not being scrollable or clickable in certain situations.